Expand packet size support to 32-bit (int) length fields
Changed packet length properties from ushort to int, raising the maximum supported packet size from 65KB to over 2GB. Updated PacketConstants, FrameBase, and PacketBase accordingly. Adjusted IS_VALID_PACKET_SIZE and removed obsolete size validation in FragmentOptions. Improved SAEA_RECEIVE_EXACTLY_ASYNC for better performance. Updated documentation and performed minor formatting fixes.

+++ src/Nalix.Common/Networking/Packets/IPacket.cs
@@ -25,7 +25,7 @@ public interface IPacket : IPacketSequenced
     /// Gets the total size, in bytes, of the serialized packet, including headers and payload.
     /// </summary>
     [SerializeIgnore]
-    ushort Length { get; }
+    int Length { get; }
 
     /// <summary>
     /// Gets the magic number that uniquely identifies the packet format or protocol.
